+++ app/assets/javascripts/compute_word_frequencies.js
@@ -2,7 +2,13 @@
 function bindWordFrequencyEvents() {
   $(document).on('change', 'input[name=block_method_switch]',
     function(event, data) {
-      $('#count_controls').toggle();
-      $('#blocks_controls').toggle();
+      $('.count_controls').toggle();
+      $('.blocks_controls').toggle();
   });
+
+  $(document).on('change', 'input[name=word_method_switch]',
+    function(event, data) {
+      $('.num_words_controls').toggle();
+      $('.inclusion_list_controls').toggle();
+    });
 }